+++ CHANGES.md
@@ -3,6 +3,8 @@ v0.3 (5/2022)
 - sample rate is a mandatory argument
 - user can input sample rate as 192k instead of 192000
 - user can specify de-emphasis time constant with -d
+- output will be resampled if an output rate is given (-R)
+- can set additional output gain with -g
 - adjusted de-emphasis to be 50 μs by default
 - adjusted audio low-pass from 15 kHz to 16.5 kHz